+++ core-service/src/main/java/com/green/core/application/controller/AdminLectureController.java
@@ -1,9 +1,9 @@
-package com.green.core.controller;
+package com.green.core.application.controller;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/majors")
-public class MajorController {
+@RequestMapping("/api/admin/lectures")
+public class AdminLectureController {
 }
