+++ core-service/src/main/java/com/green/core/application/model/lecture/LectureCreateReq.java
@@ -1,0 +1,4 @@
+package com.green.core.application.model.lecture;
+
+public class LectureCreateReq {
+}
