+++ core-service/src/main/java/com/green/core/application/service/GradeService.java
@@ -1,4 +1,4 @@
-package com.green.core.service;
+package com.green.core.application.service;
 
 import org.springframework.stereotype.Service;
