--- core-service/src/main/java/com/green/core/entity/AcademicCalandarsCache.java
@@ -1,4 +1,0 @@
-package com.green.core.entity;
-
-public class AcademicCalandarsCache {
-}
