--- core-service/src/main/java/com/green/core/entity/ProfessorCache.java
@@ -1,4 +1,0 @@
-package com.green.core.entity;
-
-public class ProfessorCache {
-}
