--- core-service/src/main/java/com/green/core/entity/StudentCache.java
@@ -1,4 +1,0 @@
-package com.green.core.entity;
-
-public class StudentCache {
-}
