+++ member-service/src/main/resources/application.yaml
@@ -24,3 +24,16 @@ spring:
   jpa:
     hibernate:
       ddl-auto: update
+  kafka:
+    bootstrap-servers: localhost:${KAFKA_PORT}
+    # [발행] student, professor 이벤트를 Outbox에 담아 보냄
+    producer:
+      key-serializer: org.apache.kafka.common.serialization.StringSerializer
+      value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
+    # [수신] Major 캐시 정보를 받기 위한 설정
+    consumer:
+      group-id: member-service-group
+      value-deserializer: org.springframework.kafka.support.serializer.JsonDeserializer
+      properties:
+        spring.json.trusted.packages: "com.green.common.model"
+
